Added Toxic chat input on player/bot death
Added class 'ToxicChatHUD' to deal with functions regarding toxic chat system. In 'ShooterHUD.cpp' in the death message method, a call is made to the ToxicChatHUD to randomly select a toxic message. This message is then used in a method call to 'AddChatLine' which sends the message to the chat.