Added Toggle Feature for Slow Chat and Filter Chat
Added Toggle Feature for Slow Chat and Filter Chat. Press Alt+1 to toggle slow chat and Alt+2 to toggle filter chat. The action binding is in ShooterPlayerController.cpp, which calls functions in ToxicChatHUD to switch the booleans for the toxic chat features. It also calls a function in ShooterHUD which then calls a function in SChatWidget to send a message to the chat stating that the slow/filter chat has been enabled/disabled.