Fixed Slow Chat Timer
Slow Chat Timer now works. Removed the blueprint and added the timer functions to ToxicChatHUD. The cause of the crashes were the result of trying to access the TimerManager from GetWorld() call, which GetWorld() was returning a null pointer since UObjects are not instantiated in the world automatically. However, the ShooterHUD that the ToxicChatHUD is initialised in is an Actor class that can call GetWorld(). Thus, this pointer now gets stored in ToxicChatHUD in the ShooterHUD constructer along with ToxicChatHUD's object initialization. This resolved the errors and the crashes.