Applied Slow Chat to incoming bot toxic messages
The slow chat feature now applies to the toxic messages received by the 'bots' when a death occurs in the game. There is no message indicating whether the timer has started for them or not, thus it may be hard to tell if this is working completely or not. I have not been able to fix the input box yet that sometimes does not allow keyboard input.